LayoutTests/fast/dom/Geolocation/geolocation-permissions-query-fingerprinting-expected.txt

This test tests the fingerprinting countermeasures implemented for querying the Geolocation API with Permissions::query().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS permissionsQueryResult is "prompt"
PASS permissionsQueryResult is "denied"
PASS successfullyParsed is true

TEST COMPLETE
